move: 메인 화면 코드를 별도의 패키지로 묶음
