+++ app/src/main/java/com/practice/hanbitlunch/screen/main/state/MainUiState.kt
@@ -1,4 +1,4 @@
-package com.practice.hanbitlunch.screen.core
+package com.practice.hanbitlunch.screen.main.state
 
 import com.hsk.ktx.date.Date
 import com.practice.database.meal.entity.MealEntity
